refactor(dpm): migrate exhibitions route to JSX format (.tsx)
